Completed an example with matching results. Starting to understand.
I am beginning to understand how to iterate over series of neutral axis
locations and curvature values to generate a P-M Interaction diagram.

This is a very rudamentary example but just working through the basics to
understand.

The example is from Design of Concrete Structure by Nilson, Darwin, Dolan
pp. 274-276. I didn't implement the whole thing just used it to check
logic in a few places.