Algorthim proof of concept for a sqaure column.
Includes the "kernel" that rotates through the section, varies the strain
profile, calculates the stress on each fiber, and then calculates the force
and moment on the section for that strain profile. This calculates the
theoretical P-M interaction curve.

Establishes basic classes for fibers and rebar.

An example of the program flow and organization.

Further development includes
    - Robust meshing utility (I know python has a package)
    - Ease of rebar pattern definition
    - Ease of shape definition
    - Plotting a 3D surface
    - Plotting multiple slices of the surface
    - Providing "filters" to apply to the theoretical curve to account
      for engineering code capacity reduction through "phi factors" and
      specific axial capacity eqauations